pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "samos"
dynamic = ["version"]                 # version comes from git tags
requires-python = ">=3.10"
dependencies = ["pydantic>=2.5", "pydantic-settings>=2.4", "PyYAML>=6.0"]
description = "SamOS – minimal runtime with soulprint, router and skills"
readme = "README.md"

[tool.setuptools.packages.find]
include = ["samos*"]
exclude = ["tests*", "examples*", "snapshots*", ".github*", "docs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[project.scripts]
samos = "samos.cli:main"

# --- setuptools_scm config ---
[tool.setuptools_scm]
# Produce clean versions from tags; no '+gHASH' in built wheels
local_scheme = "no-local-version"
# Write a build-time version file so 'import samos; samos.__version__' works in wheels
write_to = "samos/_version.py"
# HARDENING: accept only tags like v1.2.3 or v1.2.3rc1
tag_regex = "^v(?P<version>\\d+\\.\\d+\\.\\d+(?:rc\\d+)?)$"

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.11",
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "fastapi>=0.111",         # for TestClient typing
  "httpx>=0.27",            # TestClient backend
  "types-PyYAML>=6.0.12.20240808",
  "pip-audit>=2.7"          # basic security audit
]

[tool.ruff]
line-length = 120                 # relax to 120 so most lines pass
target-version = "py312"
extend-select = ["E", "F", "I"]
exclude = ["samos/_version.py"]
extend-exclude = ["tools/*"]

# Keep code strict, but don't nag about long lines in test/skills files.
[tool.ruff.per-file-ignores]
"tests/**" = ["E501"]
"samos/skills/**" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
exclude = "samos/_version.py|tools/.*"   # ← exclude tools from type-checking
